Unit tests merged into sap_solver_autodiff_test
