+++ multibody/contact_solvers/sap/BUILD.bazel
@@ -540,20 +540,6 @@ drake_cc_googletest(
 )
 
 drake_cc_googletest(
-    name = "sap_gradients_test",
-    data = ["@drake_models//:iiwa_description"],
-    deps = [                
-        "//common:find_resource",
-        "//common/test_utilities:eigen_matrix_compare",
-        "//common/test_utilities:maybe_pause_for_user",
-        "//multibody/parsing",
-        "//multibody/plant",
-        "//multibody/plant:compliant_contact_manager_tester",        
-        "//visualization:visualization_config_functions",        
-    ],
-)
-
-drake_cc_googletest(
     name = "sap_solver_test",
     deps = [
         ":sap_friction_cone_constraint",
